Tidy up filter openings
