+++ Environment_Engine/Query/FilterOpenings.cs
@@ -1,0 +1,168 @@
+﻿/*
+ * This file is part of the Buildings and Habitats object Model (BHoM)
+ * Copyright (c) 2015 - 2020, the respective contributors. All rights reserved.
+ *
+ * Each contributor holds copyright over their respective contributions.
+ * The project versioning (Git) records all such contribution source information.
+ *                                           
+ *                                                                              
+ * The BHoM is free software: you can redistribute it and/or modify         
+ * it under the terms of the GNU Lesser General Public License as published by  
+ * the Free Software Foundation, either version 3.0 of the License, or          
+ * (at your option) any later version.                                          
+ *                                                                              
+ * The BHoM is distributed in the hope that it will be useful,              
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of               
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the                 
+ * GNU Lesser General Public License for more details.                          
+ *                                                                            
+ * You should have received a copy of the GNU Lesser General Public License     
+ * along with this code. If not, see <https://www.gnu.org/licenses/lgpl-3.0.html>.      
+ */
+
+using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+using BH.oM.Environment;
+using BH.oM.Environment.Elements;
+using BH.oM.Environment.Fragments;
+using BH.oM.Base;
+
+using BH.oM.Reflection.Attributes;
+using System.ComponentModel;
+
+using BH.Engine.Base;
+
+using BH.oM.Physical.Elements;
+using BH.Engine.Geometry;
+
+using BH.oM.Geometry.SettingOut;
+
+namespace BH.Engine.Environment
+{
+    public static partial class Query
+    {
+        /***************************************************/
+        /**** Public Methods                            ****/
+        /***************************************************/
+
+        [Description("Returns a collection of Environment Openings that match the given element ID")]
+        [Input("openings", "A collection of Environment Openings")]
+        [Input("elementID", "The Element ID to filter by")]
+        [Output("openings", "A collection of Environment Opening objects that match the element ID")]
+        [PreviousVersion("3.3", "BH.Engine.Environment.Query.OpeningsByElementID(List<BH.oM.Environment.Elements.Opening>, string)")]
+        public static List<Opening> FilterOpeningsByElementID(this List<Opening> openings, string elementID)
+        {
+            List<IEnvironmentObject> envObjects = new List<IEnvironmentObject>();
+            foreach (Opening o in openings)
+                envObjects.Add(o as IEnvironmentObject);
+
+            envObjects = envObjects.ObjectsByFragment(typeof(OriginContextFragment));
+
+            envObjects = envObjects.Where(x => (x.Fragments.Where(y => y.GetType() == typeof(OriginContextFragment)).FirstOrDefault() as OriginContextFragment).ElementID == elementID).ToList();
+
+            List<Opening> rtnOpenings = new List<Opening>();
+            foreach (IEnvironmentObject o in envObjects)
+                rtnOpenings.Add(o as Opening);
+
+            return rtnOpenings;
+        }
+
+        [Description("Returns a collection of Environment Openings that match the given element ID")]
+        [Input("panels", "A collection of Environment Panels to query for openings")]
+        [Input("elementID", "The Element ID to filter by")]
+        [Output("openings", "A collection of Environment Opening objects that match the element ID")]
+        [PreviousVersion("3.3", "BH.Engine.Environment.Query.OpeningsByElementID(List<BH.oM.Environment.Elements.Panel>, string)")]
+        public static List<Opening> FilterOpeningsByElementID(this List<Panel> panels, string elementID)
+        {
+            List<Opening> allOpenings = new List<Opening>();
+            foreach (Panel p in panels)
+                allOpenings.AddRange(p.Openings);
+
+            return allOpenings.FilterOpeningsByElementID(elementID);
+        }
+
+        [Description("Returns a collection of Environment Opening that sit entirely on a given levels elevation")]
+        [Input("openings", "A collection of Environment Openings to filter")]
+        [Input("searchLevel", "The Setting Out Level to search by")]
+        [Output("openings", "A collection of Environment Openings which match the given level")]
+        [PreviousVersion("3.3", "BH.Engine.Environment.Query.OpeningsByLevel(List<BH.oM.Environment.Elements.Opening>, BH.oM.Geometry.SettingOut.Level)")]
+        public static List<Opening> FilterOpeningsByLevel(this List<Opening> openings, Level searchLevel)
+        {
+            return openings.FilterOpeningsByLevel(searchLevel.Elevation);
+        }
+
+        [Description("Returns a collection of Environment Openings that sit entirely on a given levels elevation")]
+        [Input("openings", "A collection of Environment Openings to filter")]
+        [Input("searchLevel", "The level to search by")]
+        [Output("openings", "A collection of Environment Openings which match the given level")]
+        [PreviousVersion("3.3", "BH.Engine.Environment.Query.OpeningsByLevel(List<BH.oM.Environment.Elements.Opening>, double)")]
+        public static List<Opening> FilterOpeningsByLevel(this List<Opening> openings, double searchLevel)
+        {
+            return openings.Where(x => x.MinimumLevel() == searchLevel && x.MaximumLevel() == searchLevel).ToList();
+        }
+
+        [Description("Returns a collection of Environment Openings where the maximum level of the opening matches the elevation of the given search level")]
+        [Input("openings", "A collection of Environment Openings to filter")]
+        [Input("searchLevel", "The Setting Level to search by")]
+        [Output("openings", "A collection of Environment Openings where the maximum level meets the search level")]
+        [PreviousVersion("3.3", "BH.Engine.Environment.Query.OpeningsByMaximumLevel(List<BH.oM.Environment.Elements.Opening>, BH.oM.Geometry.SettingOut.Level)")]
+        public static List<Opening> FilterOpeningsByMaximumLevel(this List<Opening> openings, Level searchLevel)
+        {
+            return openings.FilterOpeningsByMaximumLevel(searchLevel.Elevation);
+        }
+
+        [Description("Returns a collection of Environment Openings where the maximum level of the panel matches the elevation of the given search level")]
+        [Input("openings", "A collection of Environment Openings to filter")]
+        [Input("searchLevel", "The level to search by")]
+        [Output("openings", "A collection of Environment Openings where the maximum level meets the search level")]
+        [PreviousVersion("3.3", "BH.Engine.Environment.Query.OpeningsByMaximumLevel(List<BH.oM.Environment.Elements.Opening>, double)")]
+        public static List<Opening> FilterOpeningsByMaximumLevel(this List<Opening> openings, double searchLevel)
+        {
+            return openings.Where(x => x.MaximumLevel() == searchLevel).ToList();
+        }
+
+        [Description("Returns a collection of Environment Openings where the minimum level of the Opening matches the elevation of the given search level")]
+        [Input("openings", "A collection of Environment Openings to filter")]
+        [Input("searchLevel", "The Setting Out Level to search by")]
+        [Output("openings", "A collection of Environment Openings where the minimum level meets the search level")]
+        [PreviousVersion("3.3", "BH.Engine.Environment.Query.OpeningsByMinimumLevel(List<BH.oM.Environment.Elements.Opening>, BH.oM.Geometry.SettingOut.Level)")]
+        public static List<Opening> FilterOpeningsByMinimumLevel(this List<Opening> openings, Level searchLevel)
+        {
+            return openings.FilterOpeningsByMinimumLevel(searchLevel.Elevation);
+        }
+
+        [Description("Returns a collection of Environment Openings where the minimum level of the Opening matches the elevation of the given search level")]
+        [Input("openings", "A collection of Environment Opening to filter")]
+        [Input("searchLevel", "The level to search by")]
+        [Output("openings", "A collection of Environment Opening where the minimum level meets the search level")]
+        [PreviousVersion("3.3", "BH.Engine.Environment.Query.OpeningsByMinimumLevel(List<BH.oM.Environment.Elements.Opening>, double)")]
+        public static List<Opening> FilterOpeningsByMinimumLevel(this List<Opening> openings, double searchLevel)
+        {
+            return openings.Where(x => x.MinimumLevel() == searchLevel).ToList();
+        }
+
+        [Description("Returns a collection of Environment Openings that match the given opening name")]
+        [Input("openings", "A collection of Environment Openings")]
+        [Input("openingName", "The Opening Name to filter by")]
+        [Output("openings", "A collection of Environment Opening objects that match the name")]
+        [PreviousVersion("3.3", "BH.Engine.Environment.Query.OpeningsByName(List<BH.oM.Environment.Elements.Opening>, string)")]
+        public static List<Opening> OpeningsByName(this List<Opening> openings, string openingName)
+        {
+            return openings.Where(x => x.Name == openingName).ToList();
+        }
+
+        [Description("Returns a collection of Environment Openings that match a given Opening Type")]
+        [Input("openings", "A collection of Environment Openings to filter")]
+        [Input("type", "An Opening Type to filter by from the Opening Type enum")]
+        [Output("openings", "A collection of Environment Opening that match the given type")]
+        [PreviousVersion("3.3", "BH.Engine.Environment.Query.OpeningsByType(List<BH.oM.Environment.Elements.Opening>, BH.oM.Environment.Elements.OpeningType)")]
+        public static List<Opening> FilterOpeningsByType(this List<Opening> openings, OpeningType type)
+        {
+            return openings.Where(x => x.Type == type).ToList();
+        }
+    }
+}
